+++ R/learner_survivalsvm_surv_svm.R
@@ -30,7 +30,7 @@
 #' print(learner)
 #'
 #' # Define a Task
-#' task = mlr3::tsk("grace")
+#' task = mlr3::tsk("rats")
 #'
 #' # Create train and test set
 #' ids = mlr3::partition(task)