+++ R/learner_lme4_classif_glmer.R
@@ -18,8 +18,7 @@
 #' `r format_bib("bates2010lme4")`
 #'
 #' @template seealso_learner
-#' @examples
-#' if (requireNamespace("lme4", quietly = TRUE)) {
+#' @examplesIf requireNamespace("lme4", quietly = TRUE))
 #' # Define the Learner and set parameter values
 #' learner = lrn("classif.glmer",
 #'   formula = credit_risk ~ (1 | credit_history) + job + property + telephone + savings)
@@ -32,7 +31,6 @@
 #' learner$train(task)
 #'
 #' print(learner$model)
-#' }
 #' @export
 LearnerClassifGlmer = R6Class("LearnerClassifGlmer",
   inherit = LearnerClassif,