+++ R/learner_lme4_regr_lmer.R
@@ -15,8 +15,7 @@
 #' `r format_bib("bates2010lme4")`
 #'
 #' @template seealso_learner
-#' @examples
-#' if (requireNamespace("lme4", quietly = TRUE)) {
+#' @examplesIf requireNamespace("lme4", quietly = TRUE))
 #' # Define the Learner and set parameter values
 #' learner = lrn("regr.lmer", formula = cmedv ~ (1 | town))
 #'
@@ -25,7 +24,6 @@
 #'
 #' learner$train(task)
 #' print(learner$model)
-#'}
 #' @export
 LearnerRegrLmer = R6Class("LearnerRegrLmer",
   inherit = LearnerRegr,