+++ R/learner_partykit_classif_cforest.R
@@ -20,7 +20,12 @@
 #'
 #' @export
 #' @template seealso_learner
-#' @template example
+#' @examplesIf mlr3misc::require_namespaces(c("coin", "sandwich", "partykit"), quietly = TRUE)
+#' task = tsk("iris")
+#' learner = lrn("classif.cforest", ntree = 50)
+#' splits = partition(task)
+#' learner$train(task, splits$train)
+#' pred = learner$predict(task, splits$test)
 LearnerClassifCForest = R6Class("LearnerClassifCForest",
   inherit = LearnerClassif, public = list(