+++ man-roxygen/seealso_learner.R
@@ -1,6 +1,6 @@
 #' @seealso
 #' * [Dictionary][mlr3misc::Dictionary] of [Learners][mlr3::Learner]: [mlr3::mlr_learners].
-#' * `as.data.table(mlr_learners)` for a table of available [Learners][Learner] in the running session (depending on the loaded packages).
+#' * `as.data.table(mlr_learners)` for a table of available [Learners][mlr3::Learner] in the running session (depending on the loaded packages).
 #' * Chapter in the [mlr3book](https://mlr3book.mlr-org.com/): \url{https://mlr3book.mlr-org.com/basics.html#learners}
 #' * \CRANpkg{mlr3learners} for a selection of recommended learners.
 #' * \CRANpkg{mlr3cluster} for unsupervised clustering learners.