+++ NEWS.md
@@ -10,7 +10,12 @@
   and `lrn("surv.xgboost.cox")`.
 * Added early stopping and validation to catboost and lightgbm.
 * Added missing `case.depth` parameter to `rfsrc` learners.
+* `mlr3` is now in Depends instead of Imports.
+* Deprecated learner `lrn("surv.xgboost")` was now removed.
+  Use `lrn("surv.xgboost.cox")` or `lrn("surv.xgboost.aft")` instead.
 * Change xgboost default nrounds from 1 to 1000.
+* remove obliqueRSF Learner which was long superseded by aorsf
+* a lot of examples were added to the learners
 
 # mlr3extralearners 0.8.0